requirements.txt
Flask==2.3.2
groq==0.4.2
pinecone-client==3.1.0
langchain-pinecone==0.0.3
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.31.0
numpy==1.24.3
rank-bm25==0.2.2
sentence-transformers==2.2.2
spacy==3.5.3
nltk==3.8.1
scikit-learn==1.3.0
langchain==0.0.27
torch==2.0.1
transformers==4.31.0
PyPDF2
python-docx
python-pptx
tiktoken
supabase-py